+++ onedrive/src/main/scala/com/github/cloudfiles/onedrive/OneDriveFileSystem.scala
@@ -226,7 +226,7 @@ class OneDriveFileSystem(config: OneDriveConfig)
   override def updateFile(file: Model.File[String])(implicit system: ActorSystem[_]): Operation[Unit] =
     updateElement(file)
 
-  override def updateFileContent(fileID: String, size: Int, content: Source[ByteString, Any])
+  override def updateFileContent(fileID: String, size: Long, content: Source[ByteString, Any])
                                 (implicit system: ActorSystem[_]): Operation[Unit] = Operation {
     httpSender =>
       val uri = s"${itemUri(fileID)}/createUploadSession"
